pyproject.toml
[tool.poetry]
name = "slack"
version = "0.1.0"
description = ""
authors = ["Byron Kim <nocode99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
slackclient = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
